fix: `no such file or directory` when running `lx fmt`

+++ lux-cli/src/format.rs
@@ -8,7 +8,7 @@ use walkdir::WalkDir;
 // TODO: Add `PathBuf` parameter that describes what directory or file to format here.
 pub fn format() -> Result<()> {
     let project = Project::current()?.ok_or_eyre(
-        "`lux fmt` can only be executed in a rocks project! Run `lux new` to create one.",
+        "`lx fmt` can only be executed in a lux project! Run `lx new` to create one.",
     )?;
 
     let config: Config = std::fs::read_to_string("stylua.toml")
@@ -41,16 +41,18 @@ pub fn format() -> Result<()> {
 
     // Format the rockspec
 
-    let rockspec = project.root().join("project.rockspec");
+    let rockspec = project.root().join("extra.rockspec");
 
-    let formatted_code = stylua_lib::format_code(
-        &std::fs::read_to_string(&rockspec)?,
-        config,
-        None,
-        stylua_lib::OutputVerification::Full,
-    )?;
+    if rockspec.exists() {
+        let formatted_code = stylua_lib::format_code(
+            &std::fs::read_to_string(&rockspec)?,
+            config,
+            None,
+            stylua_lib::OutputVerification::Full,
+        )?;
 
-    std::fs::write(rockspec, formatted_code)?;
+        std::fs::write(rockspec, formatted_code)?;
+    }
 
     Ok(())
 }
